+++ src/HttpAssert.php
@@ -242,7 +242,7 @@ class HttpAssert
     ): Document
     {
         if (empty($expected)) {
-            return self::assertFetchedNone($strict, $contentType, $content, $message);
+            return self::assertFetchedNone($status, $contentType, $content, $message);
         }
 
         return self::assertJsonApi($status, $contentType, $content, self::STATUS_OK, $message)
@@ -383,8 +383,9 @@ class HttpAssert
         PHPUnitAssert::assertNotEmpty($id, $message ?: 'Expecting content to contain a resource id.');
 
         if (null === $expectedLocation) {
-            PHPUnitAssert::assertNull($location, 'Expecting no location header.');
+            PHPUnitAssert::assertNull($location, 'Expecting no Location header.');
         } else {
+            PHPUnitAssert::assertNotNull($location, 'Missing Location header.');
             $expectedLocation = rtrim($expectedLocation, '/') . '/' . $id;
             PHPUnitAssert::assertSame($expectedLocation, $location, $message ?: 'Unexpected Location header.');
         }
@@ -426,8 +427,9 @@ class HttpAssert
         self::assertStatusCode($status, self::STATUS_CREATED, $content, $message);
 
         if (null === $expectedLocation) {
-            PHPUnitAssert::assertNull($location, 'Expecting no location header.');
+            PHPUnitAssert::assertNull($location, 'Expecting no Location header.');
         } else {
+            PHPUnitAssert::assertNotNull($location, 'Missing Location header.');
             PHPUnitAssert::assertSame(
                 "$expectedLocation/{$expectedId}",
                 $location,
@@ -456,7 +458,13 @@ class HttpAssert
     ): void
     {
         self::assertStatusCode($status, self::STATUS_NO_CONTENT, null, $message);
-        PHPUnitAssert::assertSame($expectedLocation, $location, $message ?: 'Unexpected Location header.');
+
+        if (null === $expectedLocation) {
+            PHPUnitAssert::assertNull($location, 'Expecting no Location header.');
+        } else {
+            PHPUnitAssert::assertNotNull($location, 'Missing Location header.');
+            PHPUnitAssert::assertSame($expectedLocation, $location, $message ?: 'Unexpected Location header.');
+        }
     }
 
     /**